fixed a bug that led to an infinite loop for no-atmosphere cases
the problem was in the initialization of photons emitted by the surface. They weren't being forced to have a vertical component to their trajectory. So every once and a while they would have a perfectly horizontal path which led to them never leaving the domain, hitting the surface, of being fully absorbed