changed all reals in emission_weighting to doubles
which meant those variables had to be changed whereever else they appeared in the code. and the domain file creation tools had to be changed so that certain of those fields were double. Also, I made a slight change to how atms power was calculated to make it more consitistent with sfcpower and total_flux