bug fix and new domain creation file
most importantly I fixed a bug in the I3RC that results in a segfault when initial photon position of exactly 1.0 is converted to an index for simulations involving either irregular horizontal spacing or regular horizontal spacing but the flag indicating that is not set to true.

the file MODISreToDomain.f95 is a pieice of code that makes a domain file given some inputs defined in a .csh script. Although it was designed for the MODIS Re experiements it can easily be used or modified and used for other scenarios.