This change is largely characterized by the ability of the model to determine volume extinction coefficient from temperature, pressure, and cross section
Driver: made the physical domain a namelist varibale and removed the hardcoded-ness of the filename
OpticalProperties: added avagadro's number and the ideal gas constant as module parameters, changed read_Common to read in pressure instead of number concentration. As of right now there is no place to read in volume mixing ratios. There is only the total number conentration which is derived from the total pressure and temperature. If that feature is needed/wanted it will have to be added in later. This may have to be done for a separation of watervapor from other gasseous constituents.