This change represents removing usage of domain file list and old domain files from set up step
I replaced the calls to the domain file list and old domain files and just construct the array of domains from the common physical domain and the SSP table. A seperate step will be to address whether there needs to be an entire array of domains and weights held in memory. For now we will keep the construction of the weighting arrays as is.
     -So now there is only one common physical domain. I removed the seperate call that happened just before the main work loop and replaced all the references within the loop to the original common physical domain
     -a bunch of extraneous get_info_domain calls were commented out or removed
     -some diagnostic print statements were added to emission_weightingOLD

Changes have been tested and compared favorably to old results. A bug did have to be fixd in the ncl scripts that created the SSPtable files for the runs which was improperly converting from wavelength to frequency. It was multiplying by 10e6 instead of 1e6 or 10^6. All 4 of those scripts have been changed. This bug may have slightly impacted the runs associated with the previous commits since converting to SSPTable based.