first attempt to increase the amount of work per assignment by master
compilers, but doesn't quite work. I'm having trouble figuring out how to get the right amount of photons per batch into a group and properly loop through and keep track of the indexes containing the photons. Somehow I'm ending up with indexes below 1.

I think part of the problem is that once the loop gets to the end of the frequency array there is no mechanism to get back to the start.