model will now calculate rayleigh scattering properties if requested
requires new global attribute in SSPfile called ComponentX_calcRayl which should be set as "True" if you want to model to calculate Rayleigh scattering properties rather than read them in.

most changes were to read_SSPTable. it checks the value of calcRayl to determine whether to execute them

created a new subroutine that gets called within read_SSPTable called calc_RayleighScattering. It takes in number concentration, xsec, and lambda, and spits out extinction, singlescatteringalbedo, phaseFunctionIndex and the pahse table.

calculations in that function are based on material from the Kidder and VonderHaar book and the Liou book. I took the LG for the rayleigh phase function from the physical to optical properties tool distributed with the i3rc.