fixed the rayleigh scattering properties which were pretty far off
calculations now based solely on Kidder and vonderhaar since Liou led me astray.

model runs into an OOM and some warnings about the phase function. trying to parse it all out

created a new executable in the tools directory to print out the rayleigh properties to a text file for a set of given wavelengths.