more freeing up of memory
in the routine getOpticalComponent (or something like that) there was a souble allocation of memory...a set for the points and a set for the things the pointers ended up being assigned to. So, I just the pointers stay allocated and be assigned values directly and cut out the middle man that had no way of being deallocated.